+++ src/plugins/irc/irc-upgrade.c
@@ -384,8 +384,6 @@ irc_upgrade_read_cb (const void *pointer, void *data,
                 irc_upgrade_current_server = irc_server_search (weechat_infolist_string (infolist, "name"));
                 if (irc_upgrade_current_server)
                 {
-                    irc_upgrade_current_server->temp_server =
-                        weechat_infolist_integer (infolist, "temp_server");
                     irc_upgrade_current_server->fake_server =
                         weechat_infolist_integer (infolist, "fake_server");
                     irc_upgrade_current_server->buffer = NULL;